move basic client to /parsec folder
