+++ e2etest/cipher_test.go
@@ -9,7 +9,7 @@ import (
 	"bytes"
 	"testing"
 
-	parsec "github.com/parallaxsecond/parsec-client-go"
+	parsec "github.com/parallaxsecond/parsec-client-go/parsec"
 	"github.com/parallaxsecond/parsec-client-go/parsec/algorithm"
 )
