+++ parsec/keyinfo.go
