+++ parsec/mod.go
@@ -1,0 +1,4 @@
+// Copyright 2021 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
+
+package parsec
